refactor: separate interface
