+++ src/types/LogType.ts
@@ -1,3 +1,4 @@
+// LogType: 최종 Log 형태
 export interface LogType {
   userId: number;
   timestamp: string;
@@ -9,4 +10,23 @@ export interface LogType {
     path?: string;
     tags?: string[];
   };
+}
+
+// LoggerType: Log 내 event에 들어가는 값
+export interface LoggerType {
+  serviceName: 'drawer' | 'home' | 'search';
+  name: string;
+  message?: string;
+  path?: string;
+  tags?: string[];
+}
+
+// LogPayloadParams: 사용처에서 넣어주는 값
+export interface LogPayloadParams {
+  userId: number;
+  name: string | '';
+  serviceName?: 'drawer' | 'home' | 'search';
+  message?: string;
+  path?: string;
+  tags?: string[];
 }
